Current Best Working Code

+++ app/src/main/java/com/example/expensetracker/pojo/Entry.java
@@ -25,7 +25,7 @@ public class Entry {
         }
     }
     private String name;
-    private int value;
+    private float value;
     final private long categoryId;
     final private long subCategoryId;
     final private long paymentId;
@@ -34,7 +34,7 @@ public class Entry {
     private ArrayList<SharedUser> sharedUsersList;
     private User payer;
 
-	public Entry(long id, String name, int value, long categoryId, long subCategoryId, long paymentId, Date date, Date time, ArrayList<SharedUser> sharedUsersList) {
+	public Entry(long id, String name, float value, long categoryId, long subCategoryId, long paymentId, Date date, Date time, ArrayList<SharedUser> sharedUsersList) {
         this.id = id;
         this.name = name;
         this.value = value;
@@ -44,10 +44,10 @@ public class Entry {
         this.dateAndTime = new Date(date.getTime() + time.getTime());
         this.sharedUsersList = sharedUsersList;
     }
-    public Entry(long id, String name, int value, long categoryId, long subCategoryId, long paymentId, Date date, Date time) {
+    public Entry(long id, String name, float value, long categoryId, long subCategoryId, long paymentId, Date date, Date time) {
         this(id, name, value, categoryId, subCategoryId, paymentId, date, time, new ArrayList<>());
     }
-    public Entry(String name, int value, long categoryId, long subCategoryId, long paymentId, Date date, Date time) {
+    public Entry(String name, float value, long categoryId, long subCategoryId, long paymentId, Date date, Date time) {
         this(-1, name, value, categoryId, subCategoryId, paymentId, date, time, new ArrayList<>());
     }
 
@@ -59,7 +59,7 @@ public class Entry {
         this.name = name;
     }
 
-    public int getValue() {
+    public float getValue() {
         return value;
     }
